Added dictionaries
Dictionaries are implemented as an array of tuples. The first item
of the tuple is a hash of the key and the second item is the value
itself. The hash is translated from Ruby's hash function for string
with not modifications. When searching the dictionary for a key,
a simple loop is used, with no optimizations.